added fixes for new initialization
