+++ course-101/android/app/src/main/java/live/ditto/quickstart/tasks/models/DittoConfig.kt
@@ -3,6 +3,7 @@ package live.ditto.quickstart.tasks.models
 import androidx.annotation.Keep
 @Keep
 data class DittoConfig(
-    val endpointUrl: String,
+    val authUrl: String,
+    val websocketUrl: String,
     val appId: String,
     val authToken: String)
